Make sure not to return TRUE unless we're on the button
We use the same show_eject variable first to fetch whether the model
shows an eject button at all for the path, bailing out early if it
doesn't, and then we overwrite its value with a calculation whether the
passed-in coordinates are actually on the button.
Previously the code failed to overwrite the value of the variable when
the coordinates were not matching the button boundaries.

[endlessm/eos-shell#1341]