Made the newest tracker be the optional dependency
This allows our file manager to build with the latest tracker. Current
version of nautilus uses a bit more complex code to determine which
version to use but it also forces the depndency so to aviod messing up
the later merges from upstream, version was the only change done here.